Add a wallpaper shuffling script made in Python
My old wallpaper setter chooses a random wallpaper every time I start
the X server, which sometimes repeats images or seems to avoid
backgrounds. Of course, that's because of how randomness works. This
new script stores a shuffled stack of wallpapers, and pops out one
every time, repeating the process if the stack is empty.

I'm a bit surprised I handn't added a Python script to this repository
before.