Fix i3 not sending Firefox to its workspace
A few days ago, Debian updated the firefox-esr package, which made the
certainly unexpected move of changing the window class from "Firefox"
to "firefox-esr", for some reason. This change spawned new firefox
windows in the current workspace instead of workspace 1, as it used
to. Hopefully I don't have to do another one of this fixes.